implementations/design_10_ensemble/requirements.txt
rank_bm25>=0.2.2
faiss-cpu>=1.7.4
xgboost>=2.0.0
scikit-learn>=1.3
numpy>=1.24
httpx>=0.24
